// src/app/models/admin/dangkyhoatdongkhoangsan/pheduyettruluong/thongTinChiTietPheDuyetTruLuong.model.ts
export class OutputDKPheDuyetTruLuongKSModel {
  iddkpheduyettruluong: string;
  idgiayphep: string;
  tenbaocaothamdo: string;
  ngaylapbaocao: string;
  donvituvan: string;
  idhoso: string;
  sogiayphep: string;
  ngaycapphep: Date;
  coquancapphep: string;
}
export class InputDKPheDuyetTruLuongKSModel {
  iddkpheduyettruluong: string;
  idgiayphep: string;
  tenbaocaothamdo: string;
  ngaylapbaocao: string;
  donvituvan: string;
  idhoso: string;
  sogiayphep: string;
}